Split into "spac-core" and "spac-xml" projects (mostly)
