+++ core/src/main/scala/io/dylemma/spac/Splitter.scala
@@ -1,40 +1,13 @@
 package io.dylemma.spac
 
-import javax.xml.namespace.QName
-import javax.xml.stream.events.{StartElement, XMLEvent}
-
 import io.dylemma.spac.core.Format
-import io.dylemma.spac.handlers.{StackFormatSplitterHandler, SplitOnMatchHandler, XMLContextSplitterHandler}
+import io.dylemma.spac.handlers.{SplitOnMatchHandler, StackFormatSplitterHandler}
+import scala.language.higherKinds
 
 import scala.util.Try
 
 trait Splitter[In, +Context] {
 	def through[Out](joiner: Context => HandlerFactory[In, Out]): Transformer[In, Out]
-}
-
-trait XmlSplitter[+Context] extends Splitter[XMLEvent, Context] { splitterSelf =>
-
-	def as[Out](implicit parser: Context => HandlerFactory[XMLEvent, Out]) = through(parser)
-	def attr(name: QName) = through(Parser.forMandatoryAttribute(name))
-	def attr(name: String) = through(Parser.forMandatoryAttribute(name))
-	def asText = through(Parser.forText)
-
-	def asListOf[Out](implicit parser: Context => HandlerFactory[XMLEvent, Out]) = as[Out].parseToList
-
-	object first {
-		def apply[Out](implicit parser: Context => HandlerFactory[XMLEvent, Out]) = through(parser).parseFirst
-		def attr(name: QName) = through(Parser.forMandatoryAttribute(name)).parseFirst
-		def attr(name: String) = through(Parser.forMandatoryAttribute(name)).parseFirst
-		def asText = through(Parser.forText).parseFirst
-	}
-
-	object firstOption {
-		def apply[Out](implicit parser: Context => HandlerFactory[XMLEvent, Try[Out]]) = through(parser).parseFirstOption
-		def attr(name: QName) = through(Parser.forMandatoryAttribute(name)).parseFirstOption
-		def attr(name: String) = through(Parser.forMandatoryAttribute(name)).parseFirstOption
-		def asText = through(Parser.forText).parseFirstOption
-	}
-
 }
 
 class StackedFormatSplitter[Event, StackElem, Context](
@@ -50,24 +23,15 @@ class StackedFormatSplitter[Event, StackElem, Context](
 	def as[Out](implicit joiner: Context => HandlerFactory[Event, Try[Out]]) = through(joiner)
 }
 
+trait SplitterApply[StackElem, Instance[+_]] {
+	def apply[Context](matcher: ContextMatcher[StackElem, Context]): Instance[Context]
+}
 
 object Splitter {
 
-	def apply[Context](matcher: ContextMatcher[StartElement, Context]): XmlSplitter[Context] = new XmlSplitter[Context] { self =>
-		def through[P](joiner: Context => HandlerFactory[XMLEvent, P]): Transformer[XMLEvent, P] = {
-			new Transformer[XMLEvent, P] {
-				def makeHandler[Out](next: Handler[P, Out]): Handler[XMLEvent, Out] = {
-					new XMLContextSplitterHandler(matcher, joiner, next)
-				}
-				override def toString = s"$self{ $joiner }"
-			}
-		}
-		override def toString = s"Splitter($matcher)"
+	def apply[StackElem, Context, Instance[+_]](matcher: ContextMatcher[StackElem, Context])(implicit sa: SplitterApply[StackElem, Instance]): Instance[Context] = {
+		sa(matcher)
 	}
-
-	def generic[Event, StackElem, Context](matcher: ContextMatcher[StackElem, Context])(
-		implicit c: Format.Aux[Event, StackElem]
-	) = ???
 
 	def splitOnMatch[In, Context](matcher: PartialFunction[In, Context]): Splitter[In, Context] = {
 		new Splitter[In, Context] {self =>
