+++ core/src/main/scala/io/dylemma/spac/handlers/SequencedInStackHandler.scala
@@ -104,37 +104,3 @@ class SequencedInStackHandler[In: Stackable, T1, T2](handler1: Handler[In, T1], 
 		handler2 -> replayStack(handler2)
 	}
 }
-
-/*
-object SequencedInStackHandler {
-	def main(args: Array[String]): Unit = {
-		import io.dylemma.spac._
-		import javax.xml.stream.events.XMLEvent
-
-		val xml = """<parent>
-			<id>Dylan</id>
-			<tag>Author</tag>
-			<message>Hello</message>
-			<message>Goodbye</message>
-			<message>Have fun</message>
-		</parent>"""
-
-		case class Entity(id: String, tag: String)
-
-		val printer = Transformer.SideEffect[XMLEvent](e => println(s"  event [$e]"))
-
-		val transformer = for {
-			id <- Splitter(* \ "id").first.asText.followedByStream
-			_ = println("Got the id")
-			tag <- Splitter(* \ "tag").first.asText.followedByStream
-			entity = { println("Got the 'entity'"); Entity(id, tag) }
-			msg <- Splitter(* \ "message").asText
-		} yield s"Message from $entity: $msg"
-
-
-		val result = (printer >> transformer).consumeForEach(println).consume(xml)
-		println(s"Got result:\n  $result")
-
-	}
-}
-*/
