+++ core/src/main/scala/io/dylemma/spac/package.scala
@@ -3,11 +3,7 @@ package io.dylemma
 import io.dylemma.spac.syntax._
 
 package object spac
-	extends TransformerSyntax
-	with ConsumerSyntax
-	with ContextMatcherSyntax
+	extends ConsumerSyntax
 {
-	object TransformerSyntax extends TransformerSyntax
 	object ConsumerSyntax extends ConsumerSyntax
-	object ContextMatcherSyntax extends ContextMatcherSyntax
 }
