+++ examples/src/main/scala/io/dylemma/xml/example/Example3_EntryList.scala
@@ -1,6 +1,7 @@
 package io.dylemma.xml.example
 
 import io.dylemma.spac._
+import io.dylemma.spac.xml._
 
 /**
  * Created by dylan on 11/24/2015.
