+++ examples/src/main/scala/io/dylemma/xml/example/Example4_ParserCombine.scala
@@ -4,6 +4,7 @@ import java.text.SimpleDateFormat
 import java.util.Date
 
 import io.dylemma.spac._
+import io.dylemma.spac.xml._
 
 /**
  * Created by dylan on 10/11/2015.
