+++ examples/src/main/scala/io/dylemma/xml/example/Example_FromReadme.scala
@@ -3,6 +3,7 @@ package io.dylemma.xml.example
 import javax.xml.stream.events.XMLEvent
 
 import io.dylemma.spac._
+import io.dylemma.spac.xml._
 import org.joda.time.LocalDate
 import org.joda.time.format.DateTimeFormat
