+++ xml/src/main/scala/io/dylemma/spac/xml/handlers/MandatoryAttributeHandler.scala
@@ -1,11 +1,10 @@
-package io.dylemma.spac.handlers
+package io.dylemma.spac.xml.handlers
 
 import javax.xml.namespace.QName
 import javax.xml.stream.events.XMLEvent
 
 import io.dylemma.spac.Handler
-
-import scala.util.{Failure, Success, Try}
+import io.dylemma.spac.handlers.{FinishOnError, ManualFinish}
 
 class MandatoryAttributeHandler(name: QName)
 	extends Handler[XMLEvent, String]
